+++ changelog.md
@@ -1,3 +1,5 @@
 # Changelog
 
 - Reverted Phantom spawn attempt frequency change back to 1-2 minutes
+- Dragon should no longer bug out or path weirdly
+- Beehive spawn rates are doubled